# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IT Asset Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  general_user: "general_user",
  support_staff: "support_staff",
  administrator: "administrator",
} as const;

export type AssetStatus = (typeof AssetStatus)[keyof typeof AssetStatus];

export const AssetStatus = {
  available: "available",
  assigned: "assigned",
  in_maintenance: "in_maintenance",
  retired: "retired",
  lost: "lost",
} as const;

export type AssetCategory = (typeof AssetCategory)[keyof typeof AssetCategory];

export const AssetCategory = {
  laptop: "laptop",
  desktop: "desktop",
  monitor: "monitor",
  phone: "phone",
  tablet: "tablet",
  printer: "printer",
  server: "server",
  networking: "networking",
  peripheral: "peripheral",
  other: "other",
} as const;

export type TicketStatus = (typeof TicketStatus)[keyof typeof TicketStatus];

export const TicketStatus = {
  open: "open",
  in_progress: "in_progress",
  on_hold: "on_hold",
  resolved: "resolved",
  closed: "closed",
} as const;

export type TicketPriority =
  (typeof TicketPriority)[keyof typeof TicketPriority];

export const TicketPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface UserProfile {
  id: string;
  email: string;
  fullName: string;
  role: UserRole;
  department?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface Asset {
  id: string;
  assetTag: string;
  name: string;
  category: AssetCategory;
  status: AssetStatus;
  assignedTo?: UserProfile | null;
  purchaseDate?: string | null;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface AssetListResponse {
  data: Asset[];
  total: number;
  page: number;
  limit: number;
}

export interface TicketComment {
  id: string;
  ticketId: string;
  commentText: string;
  createdBy: UserProfile;
  createdAt: string;
}

export interface Ticket {
  id: string;
  title: string;
  description: string;
  status: TicketStatus;
  priority: TicketPriority;
  asset?: Asset | null;
  createdBy: UserProfile;
  assignedTo?: UserProfile | null;
  createdAt: string;
  updatedAt: string;
  closedAt?: string | null;
}

export type TicketWithComments = Ticket & {
  comments: TicketComment[];
};

export interface TicketListResponse {
  data: Ticket[];
  total: number;
  page: number;
  limit: number;
}

export interface Category {
  id: string;
  name: string;
  type: string;
  createdAt: string;
}

export interface DashboardStats {
  totalAssets: number;
  assignedAssets: number;
  availableAssets: number;
  inMaintenanceAssets: number;
  openTickets: number;
  inProgressTickets: number;
  resolvedTickets: number;
  totalUsers: number;
  recentTickets: Ticket[];
  recentAssets: Asset[];
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  email: string;
  password: string;
  fullName: string;
  department?: string;
}

export interface AuthResponse {
  user: UserProfile;
  token: string;
}

export interface CreateAssetRequest {
  assetTag: string;
  name: string;
  category: AssetCategory;
  status: AssetStatus;
  assignedToId?: string | null;
  purchaseDate?: string | null;
  notes?: string | null;
}

export interface UpdateAssetRequest {
  name?: string;
  category?: AssetCategory;
  status?: AssetStatus;
  assignedToId?: string | null;
  purchaseDate?: string | null;
  notes?: string | null;
}

export interface CreateTicketRequest {
  title: string;
  description: string;
  priority: TicketPriority;
  assetId?: string | null;
}

export interface UpdateTicketRequest {
  title?: string;
  description?: string;
  status?: TicketStatus;
  priority?: TicketPriority;
  assignedToId?: string | null;
}

export interface CreateCommentRequest {
  commentText: string;
}

export interface UpdateUserRequest {
  fullName?: string;
  role?: UserRole;
  department?: string | null;
}

export interface UpdateProfileRequest {
  fullName?: string;
  department?: string | null;
}

export type GetAssetsParams = {
  status?: AssetStatus;
  category?: AssetCategory;
  assignedTo?: string;
  search?: string;
  page?: number;
  limit?: number;
};

export type GetTicketsParams = {
  status?: TicketStatus;
  priority?: TicketPriority;
  assetId?: string;
  assignedTo?: string;
  search?: string;
  page?: number;
  limit?: number;
};

export type GetUsersParams = {
  role?: UserRole;
  department?: string;
  search?: string;
};

export type GetCategoriesParams = {
  type?: string;
};