# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: IT Asset Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /assets:
    get:
      operationId: getAssets
      tags: [assets]
      summary: List all assets
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/AssetStatus"
        - name: category
          in: query
          schema:
            $ref: "#/components/schemas/AssetCategory"
        - name: assignedTo
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetListResponse"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Create a new asset (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssetRequest"
      responses:
        "201":
          description: Asset created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{id}:
    get:
      operationId: getAsset
      tags: [assets]
      summary: Get asset by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Asset details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateAsset
      tags: [assets]
      summary: Update an asset (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssetRequest"
      responses:
        "200":
          description: Asset updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
    delete:
      operationId: deleteAsset
      tags: [assets]
      summary: Delete an asset (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Asset deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /categories:
    get:
      operationId: getCategories
      tags: [categories]
      summary: Get all categories
      parameters:
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a category (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    put:
      operationId: updateCategory
      tags: [categories]
      summary: Update a category (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    UserRole:
      type: string
      enum:
        - general_user
        - support_staff
        - administrator
    AssetStatus:
      type: string
      enum:
        - available
        - assigned
        - in_maintenance
        - retired
        - lost
    AssetCategory:
      type: string
      enum:
        - laptop
        - desktop
        - monitor
        - phone
        - tablet
        - printer
        - server
        - networking
        - peripheral
        - other
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        department:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - fullName
        - role
        - createdAt
        - updatedAt
    Asset:
      type: object
      properties:
        id:
          type: string
        assetTag:
          type: string
        name:
          type: string
        category:
          $ref: "#/components/schemas/AssetCategory"
        status:
          $ref: "#/components/schemas/AssetStatus"
        assignedTo:
          $ref: "#/components/schemas/UserProfile"
          nullable: true
        purchaseDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - assetTag
        - name
        - category
        - status
        - createdAt
        - updatedAt
    AssetListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Asset"
        total:
          type: number
        page:
          type: number
        limit:
          type: number
      required:
        - data
        - total
        - page
        - limit
...
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - type
        - createdAt
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        fullName:
          type: string
        department:
          type: string
      required:
        - email
        - password
        - fullName
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        token:
          type: string
      required:
        - user
        - token
    CreateAssetRequest:
      type: object
      properties:
        assetTag:
          type: string
        name:
          type: string
        category:
          $ref: "#/components/schemas/AssetCategory"
        status:
          $ref: "#/components/schemas/AssetStatus"
        assignedToId:
          type: string
          nullable: true
        purchaseDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - assetTag
        - name
        - category
        - status
    UpdateAssetRequest:
      type: object
      properties:
        name:
          type: string
        category:
          $ref: "#/components/schemas/AssetCategory"
        status:
          $ref: "#/components/schemas/AssetStatus"
        assignedToId:
          type: string
          nullable: true
        purchaseDate:
          type: string
          format: date
          nullable: true
        notes:
          type: string
          nullable: true
...