# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IT Asset Management API
 * OpenAPI spec version: 0.1.0
 */
import type { AssetCategory } from "./assetCategory";
import type { AssetStatus } from "./assetStatus";
import type { UserProfile } from "./userProfile";

export interface Asset {
  id: string;
  assetTag: string;
  name: string;
  category: AssetCategory;
  status: AssetStatus;
  assignedTo?: UserProfile | null;
  purchaseDate?: Date | null;
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}