# lib/api-zod/src/generated/types/assetListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IT Asset Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Asset } from "./asset";

export interface AssetListResponse {
  data: Asset[];
  total: number;
  page: number;
  limit: number;
}