# lib/api-zod/src/generated/types/assetStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IT Asset Management API
 * OpenAPI spec version: 0.1.0
 */

export type AssetStatus = (typeof AssetStatus)[keyof typeof AssetStatus];

export const AssetStatus = {
  available: "available",
  assigned: "assigned",
  in_maintenance: "in_maintenance",
  retired: "retired",
  lost: "lost",
} as const;