# lib/api-zod/src/generated/types/userRole.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IT Asset Management API
 * OpenAPI spec version: 0.1.0
 */

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  general_user: "general_user",
  support_staff: "support_staff",
  administrator: "administrator",
} as const;